split flake.go into flake_plan.go and flake_input.go
